+++ src/apteco/variables.py
@@ -280,6 +280,18 @@ class TextVariable(Variable):
     def between(self, start, end, *, include=True, label=None):
         normalized_start = normalize_string_value(start, single_value_error_msg_text)
         normalized_end = normalize_string_value(end, single_value_error_msg_text)
+
+        if normalized_start.lower() > normalized_end.lower():
+            if not normalized_start > normalized_end:
+                # inputted values were in correct order,
+                # but the order is switched when lower-cased
+                raise ValueError(
+                    f"`start` must come before `end`,"
+                    f" but '{normalized_start}' comes after '{normalized_end}'"
+                    f" when compared case-insensitively."
+                )
+            raise ValueError("`start` must come before `end`")
+
         return TextClause(
             self,
             [f'>="{normalized_start}" - <="{normalized_end}"'],